+++ spec/services/task_factory_spec.rb
@@ -59,4 +59,26 @@ describe TaskFactory do
     task = TaskFactory.create(klass, paper: paper, phase: phase, participants: participants)
     expect(task.participants).to eq(participants)
   end
+
+  context "roles and permissions exist" do
+    let(:journal) { create :journal }
+    let(:paper) { FactoryGirl.create(:paper, journal: journal) }
+    let(:phase) { FactoryGirl.create(:phase, paper: paper) }
+    let(:klass) { PlosBilling::BillingTask }
+    let(:journal_task_type) do
+      journal.journal_task_types.find_by(kind: klass.to_s)
+    end
+    let!(:expected_permissions) do
+      [:view, :edit].map do |action|
+        Permission.ensure_exists(action, applies_to: klass)
+      end
+    end
+
+    it "Sets default permissions from the journal_task_type" do
+      task_type_perms = journal_task_type.required_permissions
+      expect(task_type_perms).to include(*expected_permissions)
+      task = TaskFactory.create(klass, paper: paper, phase: phase)
+      expect(task.required_permissions).to include(*expected_permissions)
+    end
+  end
 end