+++ CHANGELOG.md
@@ -10,7 +10,9 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ### Deprecated
 ### Removed
 - Sending of email to Academic Editor for completed review
+- Sending of email to Academic Editor on paper resubmission
 ### Fixed
+- Ensure that cards created after paper creation have the correct permissions
 ### Security
 
 ## [1.4.3] - 2016-04-05